backend/requirements.txt
fastapi[standard]>=0.115.0
pydantic>=2.12.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
aiomysql>=0.2.0
cryptography>=42.0.0
openai>=1.0.0
python-dotenv>=1.0.0
uuid6>=2024.1.12
